backend/requirements.txt
absl-py==1.4.0
aiohttp==3.8.5
aiosignal==1.3.1
ale-py==0.8.1
annotated-types==0.7.0
anyio==4.2.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.7.2
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.1.0
audioread==3.0.1
AutoROM==0.4.2
AutoROM.accept-rom-license==0.6.1
Babel==2.14.0
base58==2.1.1
bcrypt==4.2.1
beautifulsoup4==4.12.2
bleach==6.1.0
blinker==1.7.0
cachetools==5.3.0
certifi==2022.12.7
cffi==1.17.1
charset-normalizer==3.1.0
click==8.1.7
cloudpickle==3.0.0
colorama==0.4.6
comm==0.2.1
contourpy==1.1.1
cors==1.0.1
crypto==1.4.1
cycler==0.12.1
debugpy==1.8.0
decorator==4.4.2
deepface==0.0.85
defusedxml==0.7.1
Django==5.0.1
django-cors-headers==4.5.0
dnspython==2.6.1
ecdsa==0.19.0
email_validator==2.2.0
executing==2.0.1
Farama-Notifications==0.0.4
fastapi==0.115.6
fastjsonschema==2.19.1
ffmpeg==1.4
ffmpeg-python==0.2.0
filelock==3.13.1
fire==0.5.0
Flask==3.0.1
flatbuffers==23.5.26
fonttools==4.43.1
fqdn==1.5.1
frozenlist==1.4.0
fsspec==2023.12.2
future==0.18.3
gast==0.4.0
gdown==5.0.1
gevent==24.10.3
google-auth==2.16.2
google-auth-oauthlib==1.2.0
google-pasta==0.2.0
greenlet==3.1.1
grpcio==1.51.3
gunicorn==21.2.0
gym==0.26.2
gym-notices==0.0.8
gymnasium==0.29.1
h11==0.14.0
h5py==3.8.0
idna==3.4
imageio==2.33.1
imageio-ffmpeg==0.4.9
importlib_resources==6.4.5
ipykernel==6.29.0
ipython==8.20.0
ipywidgets==8.1.1
isoduration==20.11.0
itsdangerous==2.1.2
jax==0.4.6
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.21.0
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.9.0
jupyter-lsp==2.2.1
jupyter_client==8.6.0
jupyter_core==5.7.1
jupyter_server==2.12.5
jupyter_server_terminals==0.5.1
jupyterlab==4.0.10
jupyterlab-widgets==3.0.9
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.2
keras==2.15.0
kiwisolver==1.4.5
lazy_loader==0.3
libclang==15.0.6.1
librosa==0.10.1
llvmlite==0.41.1
Markdown==3.4.1
MarkupSafe==2.1.2
matplotlib==3.8.0
matplotlib-inline==0.1.6
mistune==3.0.2
ml-dtypes==0.2.0
motor==3.6.0
moviepy==1.0.3
mpmath==1.3.0
msgpack==1.0.7
mtcnn==0.1.1
multidict==6.0.4
my-voice-analysis==0.7
Naked==0.1.32
nbclient==0.9.0
nbconvert==7.14.2
nbformat==5.9.2
nest-asyncio==1.5.9
networkx==3.2.1
node==1.2.2
noisereduce==3.0.0
notebook==7.0.6
notebook_shim==0.2.3
numba==0.58.1
numpy==1.23.5
oauthlib==3.2.2
odict==1.9.0
openai==0.28.0
opencv-python==4.9.0.80
opt-einsum==3.3.0
overrides==7.4.0
packaging==23.0
pandas==2.1.1
pandocfilters==1.5.0
parso==0.8.3
passlib==1.7.4
Pillow==10.0.1
platformdirs==4.1.0
plot-utils==0.6.14
plumber==1.7
pooch==1.8.0
praat-parselmouth==0.4.3
proglog==0.1.10
prometheus-client==0.19.0
prompt-toolkit==3.0.43
protobuf==4.22.1
psutil==5.9.7
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAudio==0.2.14
pycparser==2.21
pycryptodome==3.20.0
pycryptodomex==3.20.0
pydantic==2.10.3
pydantic_core==2.27.1
pygame==2.6.0
Pygments==2.17.2
pymongo==4.9.2
pyparsing==3.1.1
PySocks==1.7.1
python-dateutil==2.8.2
python-jose==3.3.0
python-json-logger==2.0.7
python-speech-features==0.6
pytz==2023.3.post1
pywin32==306
pywinpty==2.0.12
PyYAML==6.0.1
pyzmq==25.1.2
qtconsole==5.5.1
QtPy==2.4.1
referencing==0.32.1
requests==2.31.0
requests-file==2.1.0
requests-oauthlib==1.3.1
retina-face==0.0.13
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.17.1
rsa==4.9
scikit-learn==1.3.2
scipy==1.10.1
seaborn==0.13.2
Send2Trash==1.8.2
shellescape==3.8.1
Shimmy==0.2.1
six==1.16.0
sniffio==1.3.0
soundfile==0.12.1
soupsieve==2.5
soxr==0.3.7
SpeechRecognition==3.10.1
sqlparse==0.4.4
stack-data==0.6.3
starlette==0.41.3
sympy==1.12
tensorboard==2.15.1
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-intel==2.15.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.2.0
terminado==0.18.0
threadpoolctl==3.2.0
tinycss2==1.2.1
tk==0.1.0
tldextract==5.1.2
torch==2.1.2
tornado==6.4
tqdm==4.66.1
traitlets==5.14.1
types-python-dateutil==2.8.19.20240106
typing_extensions==4.12.2
tzdata==2023.3
uri-template==1.3.0
urllib3==1.26.15
uvicorn==0.32.1
vaderSentiment==3.3.2
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
Werkzeug==3.0.1
widgetsnbextension==4.0.9
wrapt==1.14.1
yarl==1.9.2
zope.component==6.0
zope.deferredimport==5.0
zope.deprecation==5.0
zope.event==5.0
zope.hookable==6.0
zope.interface==7.0.3
zope.lifecycleevent==5.0
zope.proxy==5.3
